feat: Create a framework
This is the initial commit that set up the framework for the event driver.

The core idea is to provide a pipelined event handler to handle the chores for users,
so that developers can focus on the business logic.

This framework would provide some standard handlers, including event composer, idempotency handler, retrier, etc.
Users can also write up a customized handler to accommodate their own usecase.