feat: Extensions use the local version of event-driver as dependency
It has been difficult to maintain the versions in extensions since they depend on a published version of event-driver,
which made it impossible to batch the changes in one-go.

A clear drawback of splitting the changes in multiple deployment is that users would see a published version that's
off-sync, which could lead to unexpected behaviors.

This PR updates the extensions to always use the local version of event-driver,
and added `#ListSourcesByKey` to `GCSEventStore` to make it implement the interface `EventStore`.